+++ memgpt/server/rest_api/server.py
@@ -1,18 +1,18 @@
-import asyncio
-import os
 from contextlib import asynccontextmanager
-import json
 
-from fastapi import FastAPI, HTTPException
-from fastapi.responses import StreamingResponse
-from pydantic import BaseModel
+from fastapi import FastAPI
 
-from starlette.exceptions import HTTPException as StarletteHTTPException
 from starlette.middleware.cors import CORSMiddleware
-from starlette.staticfiles import StaticFiles
 
+from memgpt.server.rest_api.agents.index import setup_agents_index_router
+from memgpt.server.rest_api.agents.command import setup_agents_command_router
+from memgpt.server.rest_api.agents.config import setup_agents_config_router
+from memgpt.server.rest_api.agents.memory import setup_agents_memory_router
+from memgpt.server.rest_api.agents.message import setup_agents_message_router
+from memgpt.server.rest_api.config.index import setup_config_index_router
 from memgpt.server.server import SyncServer
 from memgpt.server.rest_api.interface import QueuingInterface
+from memgpt.server.rest_api.static_files import mount_static_files
 
 """
 Basic REST API sitting on top of the internal MemGPT python server (SyncServer)
@@ -22,57 +22,11 @@ Start the server with:
   poetry run uvicorn server:app --reload
 """
 
-
-class CreateAgentConfig(BaseModel):
-    user_id: str
-    config: dict
+interface: QueuingInterface = QueuingInterface()
+server: SyncServer = SyncServer(default_interface=interface)
 
 
-class UserMessage(BaseModel):
-    user_id: str
-    agent_id: str
-    message: str
-    stream: bool = False
-
-
-class Command(BaseModel):
-    user_id: str
-    agent_id: str
-    command: str
-
-
-class CoreMemory(BaseModel):
-    user_id: str
-    agent_id: str
-    human: str | None = None
-    persona: str | None = None
-
-
-class SPAStaticFiles(StaticFiles):
-    async def get_response(self, path: str, scope):
-        try:
-            return await super().get_response(path, scope)
-        except (HTTPException, StarletteHTTPException) as ex:
-            if ex.status_code == 404:
-                return await super().get_response("index.html", scope)
-            else:
-                raise ex
-
-
-server: SyncServer | None = None
-interface: QueuingInterface | None = None
-
-
-@asynccontextmanager
-async def lifespan(application: FastAPI):
-    global server
-    global interface
-    interface = QueuingInterface()
-    server = SyncServer(default_interface=interface)
-    yield
-    server.save_agents()
-    server = None
-
+API_PREFIX = "/api"
 
 CORS_ORIGINS = [
     "http://localhost:4200",
@@ -83,7 +37,7 @@ CORS_ORIGINS = [
     "http://127.0.0.1:8283",
 ]
 
-app = FastAPI(lifespan=lifespan)
+app = FastAPI()
 
 app.add_middleware(
     CORSMiddleware,
@@ -92,110 +46,20 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
+# /api/agents endpoints
+app.include_router(setup_agents_command_router(server, interface), prefix=API_PREFIX)
+app.include_router(setup_agents_config_router(server, interface), prefix=API_PREFIX)
+app.include_router(setup_agents_index_router(server, interface), prefix=API_PREFIX)
+app.include_router(setup_agents_memory_router(server, interface), prefix=API_PREFIX)
+app.include_router(setup_agents_message_router(server, interface), prefix=API_PREFIX)
+# /api/config endpoints
+app.include_router(setup_config_index_router(server, interface), prefix=API_PREFIX)
+# / static files
+mount_static_files(app)
 
 
-@app.get("/agents")
-def list_agents(user_id: str):
-    interface.clear()
-    return server.list_agents(user_id=user_id)
-
-
-@app.get("/agents/memory")
-def get_agent_memory(user_id: str, agent_id: str):
-    interface.clear()
-    return server.get_agent_memory(user_id=user_id, agent_id=agent_id)
-
-
-@app.put("/agents/memory")
-def get_agent_memory(body: CoreMemory):
-    interface.clear()
-    new_memory_contents = {"persona": body.persona, "human": body.human}
-    return server.update_agent_core_memory(user_id=body.user_id, agent_id=body.agent_id, new_memory_contents=new_memory_contents)
-
-
-@app.get("/agents/config")
-def get_agent_config(user_id: str, agent_id: str):
-    interface.clear()
-    return server.get_agent_config(user_id=user_id, agent_id=agent_id)
-
-
-@app.get("/config")
-def get_server_config(user_id: str):
-    interface.clear()
-    return server.get_server_config(user_id=user_id)
-
-
-# server.create_agent
-@app.post("/agents")
-def create_agents(body: CreateAgentConfig):
-    interface.clear()
-    try:
-        agent_id = server.create_agent(user_id=body.user_id, agent_config=body.config)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"{e}")
-    return {"agent_id": agent_id}
-
-
-# server.user_message
-@app.post("/agents/message")
-async def user_message(body: UserMessage):
-    if body.stream:
-        # For streaming response
-        try:
-            # Start the generation process (similar to the non-streaming case)
-            # This should be a non-blocking call or run in a background task
-
-            # Check if server.user_message is an async function
-            if asyncio.iscoroutinefunction(server.user_message):
-                # Start the async task
-                await asyncio.create_task(server.user_message(user_id=body.user_id, agent_id=body.agent_id, message=body.message))
-            else:
-                # Run the synchronous function in a thread pool
-                loop = asyncio.get_event_loop()
-                loop.run_in_executor(None, server.user_message, body.user_id, body.agent_id, body.message)
-
-            async def formatted_message_generator():
-                async for message in interface.message_generator():
-                    formatted_message = f"data: {json.dumps(message)}\n\n"
-                    yield formatted_message
-                    await asyncio.sleep(1)
-
-            # Return the streaming response using the generator
-            return StreamingResponse(formatted_message_generator(), media_type="text/event-stream")
-        except HTTPException:
-            raise
-        except Exception as e:
-            raise HTTPException(status_code=500, detail=f"{e}")
-
-    else:
-        interface.clear()
-        try:
-            server.user_message(user_id=body.user_id, agent_id=body.agent_id, message=body.message)
-        except HTTPException:
-            raise
-        except Exception as e:
-            raise HTTPException(status_code=500, detail=f"{e}")
-        return {"messages": interface.to_list()}
-
-
-# server.run_command
-@app.post("/agents/command")
-def run_command(body: Command):
-    interface.clear()
-    try:
-        response = server.run_command(user_id=body.user_id, agent_id=body.agent_id, command=body.command)
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"{e}")
-    return {"response": response}
-
-
-app.mount(
-    "/",
-    SPAStaticFiles(
-        directory=os.path.join(os.getcwd(), "..", "static_files"),
-        html=True,
-    ),
-    name="spa-static-files",
-)
+@app.on_event("shutdown")
+def on_shutdown():
+    global server
+    server.save_agents()
+    server = None
